refactor(api): move association-visits to stats
